pyproject.toml
[tool.poetry]
name = "fastapi-tortoise-poetry-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["standard"], version = "^0.115.6"}
tortoise-orm = "^0.23.0"
aerich = "^0.8.1"
tortoise-database-url = "^0.1.4"
fastapi-cdn-host = "^0.8.3"
asynctor = "^0.6.8"

[tool.poetry.group.dev.dependencies]
tomli-w = "^1.1.0"
fast-dev-cli = "^0.11.5"
asgi-lifespan = "^2.1.0"

[tool.aerich]
tortoise_orm = "settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
